+++ CHANGELOG.md
@@ -2,8 +2,22 @@
 
 All notable changes to this project will be documented in this file.
 
-## ⭐ [v0.1.2] `WARNING: Break Changes with the previous version`
-> ⚡ **Feature** - Update lib Unwired.Models for the version 0.1.0##
+## 🎫 [v0.1.2] `WARNING: Break Changes with the previous version`
+> ⚡ **Feature** - Update lib Unwired.Models for the version 0.1.0<br>
 > ⚡ **Feature** - Alter name models for use the new models with prefix U
 
+## 🎫 [v0.1.3] `WARNING: Break Changes with the v0.1.1`
+> ⚡ **Feature** - Update lib Unwired.Models for the version 0.1.0<br>
+> ⚡ **Feature** - Alter name models for use the new models with prefix U
 
+## 🎫 [v0.1.4]
+> ⚡ **Feature** - Add ToInt extension method for Enum Value
+
+## 🎫 [v0.1.5]
+> ⚡ **Feature** - Update lib Unwired.Models for the version 0.1.4
+
+## 🎫 [v0.1.6]
+> ⚡ **Feature** - Update lib Unwired.Models for the version 0.1.5
+
+## 🎫 [v0.1.7]
+> ⚡ **Feature** - Create string extensions ToSnakeCase and ToCamelCase